Move the operation_mode enumeration to its own header
