+++ dlib/cuda/cudnn_dlibapi.h
@@ -3,9 +3,10 @@
 #ifndef DLIB_DNN_CuDNN_H_
 #define DLIB_DNN_CuDNN_H_
 
+#include <memory>
+#include "operation_mode.h"
 #ifdef DLIB_USE_CUDA
 #include "cuda_errors.h"
-#include <memory>
 #include "cuda_data_ptr.h"
 #endif // DLIB_USE_CUDA
 
@@ -13,17 +14,6 @@ namespace dlib
 {
     class tensor;
     class resizable_tensor;
-
-    // ----------------------------------------------------------------------------------------
-
-    /*!
-        This enum is used to determine the mode of operation for certain functions
-        (such as gemm and softmax) in Dlib. It specifies whether the calculation
-        should be performed based on the matrix field in nr()xnc() or if the matrix
-        should be considered in num_samples()xk(). This helps in organizing tensor
-        computations more efficiently according to the required dimensions.
-    !*/
-    enum class operation_mode { CHANNEL_WISE = 0, PLANE_WISE = 1 };
 
 #ifdef DLIB_USE_CUDA
     namespace cuda
